Refactor: create re-usable custom hook for useStoreContract
